requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
